dump: don't call rollback actions from a signal handler
ML: 1675ac8b8fdfcd54d76a5cc82bb4086d5067dcea

From: Andrew Vagin <avagin@virtuozzo.com>

We can do this, but we need to be sure that all structures
are consistent in any moment and we need to block alarm when
they are inconsistent.

I don't think that we really want to do this now. I suggest to
interrupt a current syscall if an alarm signal is triggered.

v2: print an error message before exiting

https://jira.sw.ru/browse/PSBM-44371

Cc: Andrey Ryabinin <aryabinin@virtuozzo.com>
Signed-off-by: Andrew Vagin <avagin@virtuozzo.com>
Reviewed-by: Andrey Ryabinin <aryabinin@virtuozzo.com>
Signed-off-by: Pavel Emelyanov <xemul@virtuozzo.com>
Signed-off-by: Cyrill Gorcunov <gorcunov@virtuozzo.com>