unshare: Support pidns
This one is tricky. When restoring into new pidns we have to provide
some init for it, and root task cannot be such, as it already has non
1 pid.

So for pidns unshare insert a fake entry as root with pid 1 and teach
it to be pseudo-init, i.e. just pick up all died tasks.

Signed-off-by: Pavel Emelyanov <xemul@virtuozzo.com>