build: Conditionally defined SYS_* constants
This patch changes the way the per architecture syscall-codes.h is
generated to define __NR* and SYS_* constants only if there are not
already defined.

This way this file could be included in the C files calling syscall()
to ensure the SYS_* constant is defined even if the C library is too
old.

The file syscall-codes.h should be included after the C library header
files.

Signed-off-by: Laurent Dufour <ldufour@linux.vnet.ibm.com>
Acked-by: Cyrill Gorcunov <gorcunov@openvz.org>
Signed-off-by: Pavel Emelyanov <xemul@virtuozzo.com>