cg: Add ability to dump predefined controllers only
When we dump container's cgroups we rely on the fact
that there is no external cgroups created on the
node which we don't know about (for this sake
we even prohibit mounting new cgroups from inside
of containers).

Still node's admin may create any additional cgroups
for own purpose unrelated to container and in such
case we will carry these cgroups inside dumped image
trying to restore them later. Moreover we're rewriting
the roots for containers so that if controller (especially
named controllers, ie cgroups mounted with "name=" options)
has path as a global root ("/") then on restore we won't
be able to recreate it properly.

For example, assume someone mounted cgroup with name=zdtmtst,
then in dump we will have

 | {
 |  "name": "name=zdtmtst",
 |  "path": "/"
 | },

Then on restore we will reqrite the root into some new
location, and the path '/' disappear because it get
overwritten with new root from command line. Of course
it's not what we wanted.

Thus for containers sake add simple filtering of cgroup
controllers to take into image files: pass their names
in option and skip anything else.

https://jira.sw.ru/browse/PSBM-46382

Signed-off-by: Cyrill Gorcunov <gorcunov@virtuozzo.com>