pcs7: scripts: spfs-release-replace.sh introduced
This script should be used to trigger actual file systems replace by SPFS
manager.

By default mount point replace processes are on hold waiting for message to
proceed with replace. The reason for this hold is that otherwise they can race
with criu on attempt to attach to processes (actual file system can be
remounted before criu released processes).

v2:
 - Updated initially missed changes

Signed-off-by: Stanislav Kinsburskiy <skinsbursky@virtuozzo.com>
Signed-off-by: Cyrill Gorcunov <gorcunov@virtuozzo.com>