pcs7: link_remap: do not create excessive links for a single file
Currently criu always creates link for each file, considered as "silly-renamed".
Even in those cases, when file is opened multiple times, it creates one new
link for each occurrence.
On restore only one link file is used for all the file descriptors, sharing
this path. And only this link is removed.
The rest of the links are forgotten and abandoned on file system.
This patch fixed this mess by creating the link only iff it's unique. Or
skipping this job otherwise.

Signed-off-by: Stanislav Kinsburskiy <skinsbursky@virtuozzo.com>
Signed-off-by: Cyrill Gorcunov <gorcunov@virtuozzo.com>