vz7: files-reg: migrate spfs unlinked file as "link remap" objects
SPFS unlinked files is a bit more tricky thing, than regular ones.
Such files have to:

1) Exist after restore.

The reason if FIFO: when it's removed from FUSE fs, it's not
accesible as file any longer (this is the way, how libfuse works; the reason
is that FIFO doesn't have "open" method in CRIU driver, and thus libfuse is
not able to hold FIFO open reference and thus can't rename file upon deletion
if opened). Also, SPFS is used for NFS only yet, and it's more natural for NFS
silly-remaned files (they are accessible on disk, while opened).

2) Has some "mark", that they are remapped.

This is required to be able to replace SPFS file by real one later after
restore.
SPFS uses extended attribute for this mark. For each path is tries to find an
attribute and if found, then file request is send to link file (and not the
original one). This allows to serve requests, when real file system was
mounted and SPFS was switched to it.
Due to this logic, link file have to available to _any_ file system, including
the fake one with created files used for restore. Or any access with result in
ENOENT.

Thus, on SPFS unlinked file restore both actual unlinked file and it's link
have to be created and _kept_ available.
This logic significantly differs from current one for "link remapped" files,
which deleted actual unlinked file after open and removes the link, once all
the files were opened.
Because of this, one more "link remap" type was introduced to be server
differently.

Feature: SPFS

Signed-off-by: Stanislav Kinsburskiy <skinsbursky@virtuozzo.com>
Signed-off-by: Cyrill Gorcunov <gorcunov@virtuozzo.com>