pcs7: mount: Corrent sibling path when parent and child have the same mountpoint
When m and pa are mounted in the same directory, rpath is initialized
to empty string (rpath[0] points to '\0'). In this case snprintf() at
the bottom of function makes path have '/' at the end.

In further, function validate_shared(), which uses the result, calls
find_shared_peer() and fails to find a peer.

https://jira.sw.ru/browse/PSBM-52346

Signed-off-by: Kirill Tkhai <ktkhai@virtuozzo.com>
Acked-by: Andrei Vagin <avagin@virtuozzo.com>
Signed-off-by: Cyrill Gorcunov <gorcunov@virtuozzo.com>