pcs7: Fix seize_cgroup_tree error code collide
Our -1 is a general code for error code, but I overlooked
that EPERM opcode is exactly equal to it, so when error
happened in seize_cgroup_tree we might end up waiting
for timeout to trigger instead of exiting immediately.

Thus use EAGAIN opcode to notify the caller that we
need to re-run freeze procedure.

Signed-off-by: Cyrill Gorcunov <gorcunov@virtuozzo.com>