vz7: Don't use executable stacks in @syscall_ip
When reporting /proc/pid/[s]maps for a process
the kernel may adjust VMA address with PAGE_SIZE
if it is a stack memrory. In turn when collecting
VMAs we do adjust such address and remember it in
VMA list.

Then, when we need to run parasite syscall with
some arguments we're walking over all VMAs and
search for one which fits in size. If the VMA
we've found has a guard page we may occasionally
try to use it and kernel won't allow us to do so.
Thus skip this page when needed.

The problem get triggered by maps05 test

https://jira.sw.ru/browse/PSBM-55989

Signed-off-by: Cyrill Gorcunov <gorcunov@virtuozzo.com>