pcs7: mnt -- Fetch VEID from cgroup for migration sake
When migrating to the other hosts we don't have VEID environment variable set,
so we either have to modify all tools (vzmigrate+p.haul+criu) to support VEID
passing via RPC call or take a simplier solution and fetch VEID from cgroup
configuration. The last one seems to take less effords so stick with it for a while.

https://jira.sw.ru/browse/PSBM-52372

Signed-off-by: Cyrill Gorcunov <gorcunov@virtuozzo.com>