pcs7: seccomp -- Switch to prctl interface if syscall is not implemented
__NR_seccomp is wired in with 48dc92b9fc3926844257316e75ba11eb5c742b2c
(v3.16-rc5-26-g48dc92b) in vanilla kernel. So lets use prctl interface
if running on older kernels.

Note that prctl interface lack of SECCOMP_FILTER_FLAG_TSYNC on which
vanilla criu relies (yielding an error in case if some of thread's
seccomp filter is not inherited), in turn here we simply assume
that filters will be inherited and restore the filter for thread
leader only.

https://jira.sw.ru/browse/PSBM-55593

Signed-off-by: Cyrill Gorcunov <gorcunov@virtuozzo.com>