action-scripts: Add "pre-resume" stage
The main idea is to be able to operate with container
at the moment where its processess and resources are
already restored but the processes are not yet in
running state, ie just before we kick them.

Beside the need of tuning up beancounters (which is vz7
specific feature) this might be useful to make some
additional debug tests from the script.

We can't reuse ACT_POST_RESUME action or move it because
we can kill the restored processes in post-resume stage
and resume them on source side as avagin@ explained.

https://jira.sw.ru/browse/PSBM-58742

Signed-off-by: Cyrill Gorcunov <gorcunov@virtuozzo.com>