vz7: pipes -- Restore pipe size via userns call
In vanilla kernels starting from v4.4-6172-g759c011
pipes allocation is limited per user-namespace, and
what is more interesting is that if hard limit is
zero but softlimit is not, then it is possible to
create new pipes with one buffer but setting its
size if not allowed until we're having CAP_SYS_RESOURCE
and CAP_SYS_ADMIN caps granted.

So to be able to restore this resource we need
to use userns daemon.

https://jira.sw.ru/browse/PSBM-59829

Signed-off-by: Cyrill Gorcunov <gorcunov@virtuozzo.com>