vz7: tty -- Don't fail if @mnt_id is not present in images
When allocating fake descriptos (say to setup control terminal)
in old images mnt_id is not present and we can't figure out
where it to be opened.

Previously we've been assuming that /dev/pts and /dev/ptmx
are standart pathes for such case, and they are for most
cases but better to lookup devpts mount point and use its
path instead. This is more portable.

https://jira.sw.ru/browse/PSBM-60606

Signed-off-by: Cyrill Gorcunov <gorcunov@virtuozzo.com>