vz7: cgroup -- Rename net_cls to net_cls,net_prio on the fly
When migrating from vz7-u3 to vz7-u4 we discovered that previously
there was only one controller for network "net_cls" which is mounted as

	/sys/fs/cgroup/net_cls

In turn on vz7-u4 systemd mounts net_cls together with net_prio controllers
into one directory

	/sys/fs/cgroup/net_cls,net_prio

So that in image we have only net_cls reference and when trying to restore container we get

    | (00.015114) cg: Skip restoring properties on cgroup dir devices//machine.slice/3e9ccc3d-229a-4872-b91c-a8d33424f4dd
    | (00.015122) cg:         Making controller dir .criu.cgyard.FYct5n/net_cls (net_cls)
    | (00.015161) Error (criu/cgroup.c:1642): cg:     Can't mount controller dir .criu.cgyard.FYct5n/net_cls: Device or resource busy

There are a few possible ways to fix it

 1) use crit tool and decode/fix/encode images before starting
   restore: very flexible but time consuming

 2) introduce some new "map" option and provide mapping from
   command line: suitable but increase code base too much

 3) implement intelligent controllers merging depending on
   node configuration: preferred way but requires a way
   more time to implement

Thus for a while here is a fast fix: we rename net_cls controller
into net_cls,net_prio on the fly to not block tests. Need to implement
(3) after all.

https://jira.sw.ru/browse/PSBM-60606
https://jira.sw.ru/browse/PSBM-61459
https://jira.sw.ru/browse/PSBM-61851

NB: Intelligent controllers merging is becoming more tempting.

Signed-off-by: Cyrill Gorcunov <gorcunov@virtuozzo.com>