vz7: Zap file size checking for "atop"
If some program has called for acct(2) and setup file
for accounting, the kernel continue wirting into the
file even upon container stop:

do_exit
 ...
 acct_process
  slow_acct_process
   do_acct_process
    ...
    __kernel_write(file, (char *)&ac, sizeof(acct_t), &pos);

For example "atop" daemon is known to doing so, and
on restore we hit a problem

 | 216: Error (criu/files-reg.c:1734): File tmp/atop.d/atop.acct has bad size 123520 (expect 121984)

because we wrote file size at moment of container stop
but kernel append data on exit.

Currently we can't handle it without modifying kernel
and criu itself, thus make a workaround and check if
the file we're dumping belong to "atop" utility
so disable file checking on it.

Note the file size cheking on itself is rather a simple
test to make sure that FS has not been modified between
c/r cycles, and in most situations it is not critical
to have such test inplace.

https://jira.sw.ru/browse/PSBM-61459

Signed-off-by: Cyrill Gorcunov <gorcunov@virtuozzo.com>

rebase to 3.15 mainstream:
move zap check to validation callbacks

Signed-off-by: Pavel Tikhomirov <ptikhomirov@virtuozzo.com>

Feature: Miscellaneous