vz7: files -- Select proper master fd when collecting fd
When choosing the master file which gonna be sending file
descriptor to the children we must not only look into
their PIDs but consider process tree relations, in particular
the child of a process might be choosen as a master and
epoll restore will fail because target files are simply
not present in child tree.

|  31964  31964  31964       epoll
|    585  31964  31964           epoll
|    586  31964  31964           epoll
|...
| (04.797121)    585: Error (criu/eventpoll.c:180): epoll: Unexpected state for tfd (id 0 fd 8)

That's because the target files are blong to 31964 and not
present in child 585, but because PID wrapp happened it
has been chosen as a leader which is of course wrong.

https://jira.sw.ru/browse/PSBM-63355

Signed-off-by: Cyrill Gorcunov <gorcunov@virtuozzo.com>