ns: Always start usernsd
We need a parent for pid_ns helpers. This can't be criu task,
as this introduces circular dependencies. So choose usernsd
for that, as we create it almost always.

Signed-off-by: Kirill Tkhai <ktkhai@virtuozzo.com>
Signed-off-by: Andrei Vagin <avagin@virtuozzo.com>

rebase: leave usernsd_pid static, also should be noted that now we are
switching netns without usernsd but it looks fine

Signed-off-by: Pavel Tikhomirov <ptikhomirov@virtuozzo.com>

rebase to mainstream 3.18:
fix other new place which uses NS_FD_OFF in userns call and add a
comment for it

https://jira.sw.ru/browse/PSBM-144416
Signed-off-by: Pavel Tikhomirov <ptikhomirov@virtuozzo.com>

Feature: Pid & User namespaces