vz7: fs: Add new argument to fstype::parse() and use it in binfmt_misc
kdat.has_binfmt_misc has two meaning. On dump it means, that
there is a binfmt_misc partition during the dumped list, and
it's not need to mount temporary binfmt_misc partition if so.

On restore it means that there is a binfmt_misc partition
in mount image (and it's not need to add a tmp partition
if there is registered entries in binfmt-misc.img).

But parse is called on dump and restore both, so it sets
kdat.has_binfmt_misc = true, when there is a binfmt_misc
partition on host.

The patch solves this problem and adds for_dump argument
to parse method. Now we will set kdat.has_binfmt_misc = true,
only if parse is called on dump, and don't do that on restore.

https://travis-ci.org/tkhai/criu/builds/231219579

Signed-off-by: Kirill Tkhai <ktkhai@virtuozzo.com>
Signed-off-by: Cyrill Gorcunov <gorcunov@virtuozzo.com>