vz7: service: allow to execute page-server as a child process
In this case we can wait it and get an exit code.

For example, it will be useful for p.haul where one connection
is used several times, so we need a way how to understand  that
page-server exited unexpectedly.

v2: don't write ps_info if a start descriptor isn't set

Signed-off-by: Andrei Vagin <avagin@virtuozzo.com>
[peeked as is, ommitting lazy-pages vanilla's commmit
1961689689f0 ("lazy-pages: fix lazy dump")]
Signed-off-by: Dmitry Safonov <dsafonov@virtuozzo.com>