pie/vdso: return back ELF header mismatch error
I've deleted it previously by the reason that I searched
vdso vma in [vdso/vvar] vma's pair by magic header.
So, I needed to suppress this error.

From that moment, I've reworked how 32-bit vdso is parsed
and now we don't need to search it, even more: we parse it
only once in the criu helper.

Signed-off-by: Dmitry Safonov <dsafonov@virtuozzo.com>