vz7: spfs: pass both "mountpoint" and "ns_mountpoint" paths to manager
Now, these two paths are different even for root mount namespace.
The first one (mountpoint) is the path, where the mount has to be mounted to
restore it.
The second one (ns_mountpoint) is the path, where mount will be located
_within__ container, and, thus, this pass has to be used to place real mount
to.
One more difference: "root" is not required anymore. Originally, path to mount
SPFS was constructed as follows:

SPFS_mnt_path = CT_root + mi->ns_mountpoint

Now mounts are created elsewhere and full path is defined in mi->mountpoint.

Signed-off-by: Stanislav Kinsburskiy <skinsbursky@virtuozzo.com>
https://jira.sw.ru/browse/PSBM-66943
Signed-off-by: Dmitry Safonov <dsafonov@virtuozzo.com>