vz7: vdso: Don't read pagemap or parse symtable
We can save some syscalls for *each* dumpee if we don't
open()/seek()/read()/close() /proc/pid/pagemap for each
dumpee and even don't use parasite to parse symtable if
pagemap is unavailable.

This is a backport of vanilla CRIU's patch and it should
be dropped in favour of it.
I didn't backport kdat part of test, because on v3.0 criu
kdat is not runtime-cached file, but tested each criu launch.
So, as we know that vzkernel preserve correctly "[vdso]" hint,
no need to test it each time.

https://jira.sw.ru/browse/PSBM-67017

[Drop this after CRIU ~v3.3 in favour of vanila patches]
Signed-off-by: Dmitry Safonov <dsafonov@virtuozzo.com>