cr-service: Init page_ids laterly
As pre-dump, dump and page-server may run as a commands during
one service session, don't set it up early in main() - as
on the next command the value of page_ids will be preserved.
Rather do it lately during initialization of (pre-)dump/pageserver.
As (pre-)dump/pageserver rpc commands are executed under fork(),
we will not share the counter to criu.

Signed-off-by: Dmitry Safonov <dsafonov@virtuozzo.com>