mounts: support only shared NFS mounts in inits mount namespace
The intention of this patch is to define different NFS mount cases and forbid
most ot them.
The reason behind this is that the only mount point we can migrate right now
is:
1) created in init mount namespace
2) shared (because it's default)

All the other cases are not supported by spfs manager so far and migration
attempt can lead to different issues after restore is completed by CRIU.

https://jira.sw.ru/browse/PSBM-66945

v2:
1) Added check that all bind-mounts are in the same shared group for slave
mount. We need this to make sure, that all such mounts will be propagated
correctly.

Signed-off-by: Stanislav Kinsburskiy <skinsbursky@virtuozzo.com>
Reviewed-by: Andrei Vagin <avagin@virtuozzo.com>
Signed-off-by: Dmitry Safonov <dsafonov@virtuozzo.com>