vz7: bc, restore -- Check for failcounters on error
When there is no enough memory on the host to proceed
restore procedure, the OOM killer may simply zap any
suitable process, which can be even criu itself so
without reading of dmesg output we can't figure out
why error happened.

In better case the criu remains working but with some
of subroutines failed, for such case we add scanning
of beancounters output for @failcnt value and report
ENOMEM here. While it's not reliable it still might
be useful for analisys.

https://jira.sw.ru/browse/PSBM-68868

Signed-off-by: Cyrill Gorcunov <gorcunov@virtuozzo.com>