sk-unix: Handle bindmounted dgram sockets on restore
If a socket is bindmounted one we open it early (thus mount engine
will be eable to proceed), put it into fdstore and then simply fetch
it from there.

Note it's important that only basic scenario with dgram/closed sockets
is supported by now: established, listening sockets requires a way
more code change and we will do it only when really needed.

https://jira.sw.ru/browse/PSBM-69473

Signed-off-by: Cyrill Gorcunov <gorcunov@virtuozzo.com>