vz7: image -- Add compat layer for Core entry
In 3.1 the TaskCoreEntry members tty_nr and tty_pgrp get
clashed with vz7 instance. Thus if container has been
checkpoint with vz7-u5 (3.0 series) it won't be restored
with vz7-u6 (3.4 series).

To resolve it we introduce task_core_entry_VZ730 and
core_entry_VZ730 members which get read and converted
into new format on the fly.

The patch is likely to be dropped in next major vz7
release (update 7 likely).

https://jira.sw.ru/browse/PSBM-67671

Signed-off-by: Cyrill Gorcunov <gorcunov@virtuozzo.com>