tty: Use mnt_id for first lookup of toplevel devpts
When we're searching for toplevel devpts mount into which
the toplevel /dev/ptmx device is pointing to we should use
mnt_id from mountinfo first, this is most reliable way to
detect mounting.

But on old kernels mnt_id may not be present so we should
try s_dev lookup then and here is a nit -- mointinfo keeps
superblock device in kernel format while the source s_dev
is obtained from the stat() syscall so we should use
kdev_to_odev helper to not miss device if there are many
devpts mounted on the node.

https://jira.sw.ru/browse/PSBM-78460
Signed-off-by: Cyrill Gorcunov <gorcunov@virtuozzo.com>

Feature: TTY