files-reg: Prepare for sysfs entries mode change
The kernel virtualize access to proc/sys/ entries in
lightweight way -- if entry is opened from inside of
veX then it's not allowed to be written.

Still we're dumping files in ve0 environment so the
mode for such files may no match on restore, because
restore itself is running inside veX.

 | (00.701122)    111: Error (criu/files-reg.c:1976): File proc/sys/vm/overcommit_kbytes has bad mode 0100444 (expect 0100644)
 | (00.701496)    334: Error (criu/cr-restore.c:1353): 111 exited, status=1

so i think we can simply skip such testing inside ve criu
instance since it's kernel specific.

Simply print out a warning for refernce and continue

 | (00.827328)    111: Error (criu/files-reg.c:1977): File proc/sys/vm/overcommit_kbytes has bad mode 0100444 (expect 0100644)
 | (00.827332)    111: Warn  (criu/files-reg.c:1985):      Expecting in VE environment. Ignore.

https://jira.sw.ru/browse/PSBM-80585

Signed-off-by: Cyrill Gorcunov <gorcunov@virtuozzo.com>

rebase to 3.15 mainstream:
FIXME: checking "proc/sys" path is not robust

Feature: Miscellaneous

Signed-off-by: Pavel Tikhomirov <ptikhomirov@virtuozzo.com>