vz7: netlink: request flags for netlink sockets
We need to know where a socket has a running callback or not.
If it has, we can't dump its received queue, because it
contains only a part of data, other part will be generated by the callback.

Signed-off-by: Andrew Vagin <avagin@virtuozzo.com>
Signed-off-by: Cyrill Gorcunov <gorcunov@virtuozzo.com>

rebase to mainstream 3.17.1:
fix hunk in collect_sockets which was broken in vz7-u11

Feature: Netlink packet queue

https://jira.sw.ru/browse/PSBM-144416
Signed-off-by: Pavel Tikhomirov <ptikhomirov@virtuozzo.com>