vz7: sk-queue -- Add support of ucreds in unix sockets
Checkpoint and restore ucreds on unix sockets are not
that trivial: the @pid returned by the kernel upon
message fetching is mapped to current user namespace
so it is not the same as it would be received inside
container. So we need to map real pid returned into
virtual seen from inside, but we know virtual pids
only when complete process tree is built.

Thus what we do -- queue such packets into own
queue and when process tree is build walk over
and transform former pids into virtual ones.

Feature: Ucreds

https://jira.sw.ru/browse/PSBM-61095
Signed-off-by: Cyrill Gorcunov <gorcunov@virtuozzo.com>