mem: pid-reuse -- Don't fail if no stat image present
The statistic file is written into cwd directory and
in result we can't open it back when migrating container
with pre-dump stage.

Thus we need to prepare criu and p.haul to serve stat
file on per-image-directory basis. For now to not
break existing migrations lets just test for statistics
absence and continue without pid-reuse detection.

https://jira.sw.ru/browse/PSBM-82806

Signed-off-by: Pavel Tikhomirov <ptikhomirov@virtuozzo.com>
Signed-off-by: Cyrill Gorcunov <gorcunov@virtuozzo.com>