files: Don't fail on cgroup directories mode change
When restoring cgroup directories we're keeping only controllers
the task lives in and due to vz specifics we restore missed dentries
with 0755 mode. But actually systemd subgroups may be controlled by
node operator or even modified from inside of a container.

For example one can create dentry chain and move some task into nested
dentry only but change the mode by hands on a parent dentry. This is
a very uncommon case and for now we can simply put a warning in restore
log instead of failing with error.

 | (00.773246)     88: Error (criu/files-reg.c:1979): \
 | File sys/fs/cgroup/systemd/user.slice/user-0.slice/user@0.service
 | has bad mode 040755 (expect 040750)

Or

 | (03.343479)      1: Error (criu/files-reg.c:1977): \
 | File sys/fs/cgroup/systemd has bad mode 040700 (expect 046755)

If indeed it become a critical one day we will have
to invent a way how to save complete dentry paths.

https://jira.sw.ru/browse/PSBM-83154

Signed-off-by: Cyrill Gorcunov <gorcunov@virtuozzo.com>

rebase to 3.15 mainstream:
FIXME: checking "sys/fs/cgroup" path is not robust

Feature: Miscellaneous

Signed-off-by: Pavel Tikhomirov <ptikhomirov@virtuozzo.com>