files: Fail dump if dump_one_file() fails
When dumping a process with a large number of open files,
dump_task_files_seized() processes the fds in batches.
If dump_one_file() results in an error, processing
of the current batch is stopped but the next batch (if any)
will still be fetched and the error value is overwritten.
The result is a corrupt dump image (the fdinfo file is missing
a bunch of fds) which results in restore failure.

Also close all received fds after an error
(previously the skipped ones were left open).

Signed-off-by: Andy Tucker <agtucker@google.com>
Signed-off-by: Cyrill Gorcunov <gorcunov@virtuozzo.com>