pid-reuse: temporary use stat file from cwd if has no parent stats
We will eventually switch from stats file to inventory image, as the
former is a wrong place for dump_uptime anyway. But for now to reenable
pid-reuse detection for vz7 CT's we can read dump_uptime of a previous
pre-dump from stats file in cwd, as libvzctl sets for all iterations
same workdir, and as we read it before cr_*dump_finish (before we
overwrote it).

https://jira.sw.ru/browse/PSBM-82864

Signed-off-by: Pavel Tikhomirov <ptikhomirov@virtuozzo.com>