unix: Add support for bindmounted sockets
Some unix sockets might be bindmounted (say /dev/log
bound to another place). So to handle it we need to
change the logic we open such sockets especially
because we create mount tree earlier than we start
to restore files.

Thus here what we do:

 - on dump mark such sockets with UNIX_UFLAGS__BINDMOUNT
   flag so we would distinguish them on restore;

 - collect unix sockets before creating mount tree;
   note that at this moment we able to simply gather
   this sockets into own @unix_mnt_sockets list and
   nothing more because setting up the peers and such
   happens later in that named post action procedures;

 - when we need to create a bindmount point we enter
   into unix engine and figure out if there a socket
   to bindmount over; if found we pre-allocate the
   socketpair, bind it and save inside fdstore engine;
   using socketpair is important because later we need
   both peers to restore queued data;

 - finally when we start restoring files we simply
   fetch the socket from the fdstore and use it
   directly.

All this scheme is working simply because we support
dgram standalone sockets only, adding support for
streamed sockets requires a way more engine rework
and hopefully we won't need it in near future.

https://jira.sw.ru/browse/PSBM-82616

Signed-off-by: Cyrill Gorcunov <gorcunov@gmail.com>