mount: don't lose shared options for bind mounts
When we bind-mount something as a child of shared group, child will be
created shared, so we should fix sharing flags after bind but not
before.

Bigger problem we face is: when some mount is wrongly created shared
instead of private, it's children can be duplicated, and migration may
hit container limit on number of mounts, after couple of iterations.

https://jira.sw.ru/browse/PSBM-85251

Signed-off-by: Pavel Tikhomirov <ptikhomirov@virtuozzo.com>